backend/app/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
python-dotenv
requests
pymupdf
